+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import MainButton from '@/components/MainButton/MainButton';
 import styles from './page.module.css';
 
 import { Provider } from 'react-redux';
-import store from '@/store/store';
+import store from '@/redux/store';
 import MainMenu from '@/components/MainMenu/MainMenu';
 import ChatDialog from '@/components/ChatDialog/ChatDialog';
