+++ src/components/AttachmentFilesButton/AttachmentFilesButton.module.css
@@ -4,7 +4,7 @@
   display: block;
   cursor: pointer;
   width: 40px;
-  background-image: url('../../images/paperclip.svg');
+  background-image: url('../../../public/images/paperclip.svg');
   background-size: contain;
   background-repeat: no-repeat;
   background-position: center;
